+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2021-04-26
+## [1.1.1][1.1.1] - 2021-05-01
+### Added
+- Linting with eslint
+
+### Changed
+- `var` to `const` in demo HTML
+- All source js files to comply with linting rules
+
+### Removed
+- Removed unused dev dependencies
+- Generated .css file inside dist
+- Lock file: `package-lock.json`
+
+## [1.1.0][1.1.0] - 2021-04-26
 ### Added
 - This changelog.
 
@@ -15,4 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Bower support (very few people using it).
 
+[1.1.1]: https://github.com/fegemo/bespoke-simple-overview/compare/v1.1.1...v1.1.0
 [1.1.0]: https://github.com/fegemo/bespoke-simple-overview/compare/v1.1.0...v1.0.0